+++ src/GoldyBot/assets.py
@@ -1,2 +1,5 @@
 NO_PFP_IMAGE = "https://media.discordapp.net/attachments/436201641486581762/967863279575765042/logo_v2.png"
 """"""
+
+ANIME_ONE_PUNCH_CLOCK_GIF = "https://bestanimations.com/media/clocks/1998493116funny-alarm-clock-animated-gif-3.gif"
+""""""