+++ src/GoldyBot/errors.py
@@ -1,14 +1,14 @@
 import GoldyBot
 from GoldyBot.logging import print_and_log
 
-import time
-
 class GoldyBotError(Exception):
     """The parent/master class of all the errors in Goldy Bot."""
     def __init__(self, error, no_traceback:bool=False):
         """Raise me! UwU OwO"""
         super().__init__(error)
         GoldyBot.logging.log("error", error, no_traceback)
+
+
 
 # Commands
 #-------------
@@ -24,6 +24,8 @@ class MemberHasNoPermsForCommand(GoldyBotError):
     """Raised when a member without the proper roles tries to execute a command."""
     pass
 
+
+
 # Roles
 #-----------
 class FailedToFindRole(GoldyBotError):
@@ -34,6 +36,18 @@ class FailedToFindRole(GoldyBotError):
             error = f"Failed to find role by the {option_used} '{option_value}'."
 
         super().__init__(error)
+
+
+
+# Members
+#-----------
+class FailedToFindMember(GoldyBotError):
+    """This exception is raised by ``GoldyBot.objects.Member`` when the member can't be found."""
+    def __init__(self):
+        error = "Failed to find Member! Double check you've entered the id correctly or you've passed the correct object."
+        super().__init__(error)
+
+
 
 # Modules
 #------------
@@ -61,8 +75,8 @@ class FailedToFindGuildRole(GoldyBotError):
 
 # Config Incorrect
 class ConfigIsIncorrect(GoldyBotError):
+    """Error that is raised whenever a configuration file is incorrect or missing something."""
     def __init__(self, error, what_is_incorrect:str=None):
-        """Error that is raised whenever a configuration file is incorrect or missing something."""
         super().__init__(error)
         print_and_log("info", "^ This usually means you misspelt something in the config or the layout of the config is incorrect. ^")