+++ src/GoldyBot/objects/command.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 import asyncio
 import importlib
-from typing import List
+from typing import Dict, List
 import nextcord
 from nextcord.ext import commands
 
@@ -122,9 +122,14 @@ class Command(Embeds):
 
         return_data = {}
 
+        #  Make slash command only visible to admins if it is hidden.
+        default_member_permissions = None
+        if self.is_hidden: default_member_permissions = nextcord.Permissions(permissions=8)
+        
+
         if self.in_extension == True: # Run in EXTENSION!
             exec(f"""
-@client.slash_command(name=command_name, description=help_des, guild_ids=guilds_allowed_in)
+@client.slash_command(name=command_name, description=help_des, guild_ids=guilds_allowed_in, default_member_permissions=default_member_permissions)
 async def slash_command_(interaction: Interaction{slash_command_params[0]}):
     ctx = GoldyBot.objects.slash.InteractionToCtx(interaction)
     try:
@@ -152,11 +157,12 @@ async def slash_command_(interaction: Interaction{slash_command_params[0]}):
             
             {"func":self.func, "client":GoldyBot.cache.main_cache_dict["client"], "command_name":self.command_name, "help_des":self.get_help_des(), "self":self,
             "Interaction": GoldyBot.nextcord.Interaction, "GoldyBot": GoldyBot, "asyncio":asyncio, "nextcord":nextcord, "class_":self.extension, "no_perms_embed":self.no_perms_embed, 
-            "guild_not_registered_embed":self.guild_not_registered_embed, "hidden":self.is_hidden, "guilds_allowed_in":self.guilds_allowed_in}, return_data)
+            "guild_not_registered_embed":self.guild_not_registered_embed, "hidden":self.is_hidden, "guilds_allowed_in":self.guilds_allowed_in, 
+            "default_member_permissions":default_member_permissions}, return_data)
             
         else: # Run as NORMAL command!
             exec(f"""
-@client.slash_command(name=command_name, description=help_des, guild_ids=guilds_allowed_in)
+@client.slash_command(name=command_name, description=help_des, guild_ids=guilds_allowed_in, default_member_permissions=default_member_permissions)
 async def slash_command_(interaction: Interaction{slash_command_params[0]}):
     ctx = GoldyBot.objects.slash.InteractionToCtx(interaction)
     try:
@@ -184,7 +190,8 @@ async def slash_command_(interaction: Interaction{slash_command_params[0]}):
             
             {"func":self.func, "client":GoldyBot.cache.main_cache_dict["client"], "command_name":self.command_name, "help_des":self.get_help_des(), "self":self,
             "Interaction": GoldyBot.nextcord.Interaction, "GoldyBot": GoldyBot, "asyncio":asyncio, "nextcord":nextcord, "no_perms_embed":self.no_perms_embed, 
-            "guild_not_registered_embed":self.guild_not_registered_embed, "hidden":self.is_hidden, "guilds_allowed_in":self.guilds_allowed_in}, return_data)
+            "guild_not_registered_embed":self.guild_not_registered_embed, "hidden":self.is_hidden, "guilds_allowed_in":self.guilds_allowed_in, 
+            "default_member_permissions":default_member_permissions}, return_data)
 
         GoldyBot.logging.log(f"[{MODULE_NAME}] [{self.command_name.upper()}] Slash command created!")
 
@@ -310,7 +317,7 @@ async def slash_command_(interaction: Interaction{slash_command_params[0]}):
         """Adds/updates this command object to the class."""
         self.command = command_object
 
-    def sub_command(self, command_name:str=None, required_roles:list=[], help_des:str=None, slash_options:dict={}, also_run_parent_CMD:bool=True):
+    def sub_command(self, command_name:str=None, required_roles:list=[], help_des:str=None, slash_options:Dict[str, nextcord.SlashOption]={}, also_run_parent_CMD:bool=True):
         """Create a lovely sub command from this slash command. 😀 (Only works with slash commands.)"""
 
         def decorate(func):
@@ -418,10 +425,16 @@ async def slash_command_(interaction: Interaction{slash_command_params[0]}):
                         if str(ctx.author.id) in GoldyBot.settings.BOT_DEVS:
                             return True
 
+                    # If the required roles contain 'bot_admin' and a bot admin is running the command allow the command to execute. (NEW)
+                    #----------------------------------------------------------------------------------------------------------------
+                    if "bot_admin" in self.required_roles_:
+                        if str(ctx.author.id) in goldy_config.read("admin_users"):
+                            return True
+
                     # Check if member has any of the required roles.
                     #----------------------------------------------------
                     for role_code_name in self.required_roles_:
-                        if not role_code_name == "bot_dev":
+                        if not role_code_name in ["bot_dev", "bot_admin"]:
                             role = guild_config.get_role(ctx, role_code_name)
                             if GoldyBot.objects.member.Member(ctx).has_role(role):
                                 return True