+++ src/GoldyBot/utility/misc/__init__.py
@@ -1,0 +1,4 @@
+
+def merge_dict(dict1:dict, dict2:dict) -> dict:
+    """Merges to dictionaries together."""
+    return {**dict1, **dict2}